🔨 Fixed templates for graph abstractions
1. Created a separate Jinja template for GPMA, PCSR and CSR.
2. Changed folder structure to have a template/ folder
3. Changed the datatype of value in gpma.cu from double to unsigned int

+++ seastar/graph/dynamic/gpma/gpma.cu
@@ -31,7 +31,7 @@ __inline__ __host__ __device__ void gpuAssert(cudaError_t code, const char *file
 }
 
 typedef unsigned long long KEY_TYPE;
-typedef double VALUE_TYPE;
+typedef unsigned int VALUE_TYPE;
 typedef unsigned int SIZE_TYPE;
 
 typedef thrust::device_vector<KEY_TYPE> DEV_VEC_KEY;
